backend/requirements.txt
# Flask core
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Database
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.9
flask-migrate>=4.0.5

# Authentication
flask-jwt-extended>=4.6.0
flask-bcrypt>=1.0.1

# WebSocket
python-socketio>=5.10.0
eventlet>=0.33.0

# Google OAuth
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

